4/19/2022
The changes made today in the LEGO code was that the initial turn was added so that it will align with the vertical electrophoresis pole, after grabbing the first two rings. The biggest problem that needs to be fixed, is the motors don't drive at the same speed, so the robot tries to drive off of the board.