Lego bot 4/28/2022
The claw was changed so that it picked up the rings slower, which helped it to pick up both rings every time. Now, the robot lines up with the pole perfectly, but the rings roll off the side of the pole, so next time we are going to try and open the claw slower using the same while loop used for picking up the rings.